Update switching from VS Code to Chrome heuristics
Able to use websocket with webpack so we can now open two-way communication between the Chrome extension and the node server (or node folder in webActivities). Essentially node will handle app switch context, and when user switches from VS Code to Chrome, if the page is localhost, Chrome will just reload so we can have a record of that in our data table. This for now is a decent approach considering live reload environment complication and not requiring user to manually reload the page.