filtering tabs with recording on in db
The reason to use tableData is to store a history of navigation. Filtering the db would only look like a history if the recording is turned on for all tabs. Let's say if recording turns on for all tabs, then off, and then on again for a few tabs, then filtering would only contain those few tabs. However, tableData which appends tab with recording on would still contain the history with past recorded tabs => Either doing tableData as well in IndexedDB, keeping the current tableData from chrome.storage and assist it with IndexedDB, or appending infos directly to the local webData file (need to watch out the format for this)