Mapped functionality of the draggable control
URLs such as "chrome://*" or "https://developer.chrome.com/*", etc and locally open URL like PDFs have strict security and does not allow script injection. Thus, the float control would not show up. For the majority of other websites, if the recording is on, the float will be in green color with option to stop recording. If the recording is off, the float will be in purple with option to start recording. The hide option will hide the control. Can add a small option to show it again but for now, user can use context menu to show it.